backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
qrcode==7.4.2
Pillow==10.2.0
psycopg2-binary==2.9.9
PyMuPDF==1.23.26
python-docx==1.1.0
anthropic==0.25.0
chromadb==0.4.24
sentence-transformers==2.6.1
langchain==0.1.16
langchain-community==0.0.34
langchain-anthropic==0.1.6
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
playwright==1.43.0
beautifulsoup4==4.12.3
httpx==0.27.0
azure-storage-blob==12.19.1
sentry-sdk==1.43.0